Spike out an alternate approach.
This changes how the library is imlemented significantly while keeping
the API _mostly_ the same. Instead of defining the action functions on
your controller for you we instead intercept calls to those actions via
a plug and dispatch them to the behaviour. The behavour then hits the
callbacks as required to serve the request.